+++ include/vector.h
@@ -1,0 +1,31 @@
+/************************************************************************
+* Copyright (C) 2020 by Ahmad Dajani                                    *
+*                                                                       *
+* This file is part of NOS.                                             *
+*                                                                       *
+* NOS is free software: you can redistribute it and/or modify it        *
+* under the terms of the GNU Lesser General Public License as published *
+* by the Free Software Foundation, either version 3 of the License, or  *
+* (at your option) any later version.                                   *
+*                                                                       *
+* NOS is distributed in the hope that it will be useful,                *
+* but WITHOUT ANY WARRANTY* without even the implied warranty of        *
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+* GNU Lesser General Public License for more details.                   *
+*                                                                       *
+* You should have received a copy of the GNU Lesser General Public      *
+* License along with NOS.  If not, see <http://www.gnu.org/licenses/>.  *
+************************************************************************/
+
+/*@file vector.h
+* @author Ahmad Dajani <eng.adajani@gmail.com>
+* @date 1 Jan 2021
+* @brief Interrupt vector header file
+*/
+
+#ifndef __VECTOR_H
+    #define __VECTOR_H
+
+    void setInterruptVector(int interruptNumber, void interrupt (*interruptFunctionPointer)());
+    void interrupt (*getInterruptVector (int interruptNumber))(void);
+#endif